cms: split software / hardware header files
